minor issues fixed
